reorganize class structure
